fixed redirection bug & organized the project
