+++ app/imports/ui/deprecated/ListStuffAdmin.jsx
