+++ app/imports/ui/deprecated/NavBar.jsx
