+++ app/imports/ui/deprecated/StuffItem.jsx
