+++ app/imports/ui/deprecated/StuffItemAdmin.jsx
