+++ app/imports/ui/pages/UserProfile.jsx
@@ -1,10 +1,11 @@
 import React from "react";
 import { Meteor } from "meteor/meteor";
-import { Container, Header, Loader } from "semantic-ui-react";
+import { Header, Loader } from "semantic-ui-react";
 import { withTracker } from "meteor/react-meteor-data";
 import PropTypes from "prop-types";
 import Users from "../components/Users";
 import { User } from "../../api/user/User";
+import { Redirect } from "react-router-dom";
 
 /** Renders a table containing all of the Stuff documents. Use <StuffItem> to render each row. */
 class UserProfile extends React.Component {
@@ -18,7 +19,9 @@ class UserProfile extends React.Component {
 
   // Render the page once subscriptions have been received.
   renderPage() {
-    return (
+    return this.props.users?.length === 0 ? (
+      <Redirect to={{ pathname: "/editUserProfile/new" }} />
+    ) : (
       <div className="flex-column">
         <Header as="h2" style={{ textAlign: "center" }}>
           My Profile
@@ -37,14 +40,9 @@ UserProfile.propTypes = {
   ready: PropTypes.bool.isRequired,
 };
 
-// withTracker connects Meteor data to React components. https://guide.meteor.com/react.html#using-withTracker
 export default withTracker(() => {
-  // Get the documentID from the URL field. See imports/ui/layouts/App.jsx for the route containing :_id.
-  // Get access to Stuff documents.
   const subscription = Meteor.subscribe(User.userPublicationName);
-  // Determine if the subscription is ready
   const ready = subscription.ready();
-  // Get the Stuff documents
   const users = User.collection.find({}).fetch();
   return {
     users,
